move create_comet_physical_fun back into core and make scalar funcs public
