+++ native/spark-expr/src/scalar_funcs/hex.rs
@@ -62,7 +62,7 @@ fn hex_bytes<T: AsRef<[u8]>>(bytes: T) -> Result<String, std::fmt::Error> {
     Ok(hex_string)
 }
 
-pub(super) fn spark_hex(args: &[ColumnarValue]) -> Result<ColumnarValue, DataFusionError> {
+pub fn spark_hex(args: &[ColumnarValue]) -> Result<ColumnarValue, DataFusionError> {
     if args.len() != 1 {
         return Err(DataFusionError::Internal(
             "hex expects exactly one argument".to_string(),
